chore: moved crypto and flattened plist
