+++ internal/common/crypto/hash.go
