+++ internal/common/plist/normalize_payload.go
