+++ internal/common/plist/shared.go
