+++ internal/common/plist/validate.go
