+++ internal/services/macos_configuration_profile_plist/resource_constructor.go
@@ -9,8 +9,8 @@ import (
 	"strings"
 
 	"github.com/deploymenttheory/go-api-sdk-jamfpro/sdk/jamfpro"
-	helpers "github.com/deploymenttheory/terraform-provider-jamfpro/internal/common/configuration_profiles/plist"
 	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/common/constructors"
+	helpers "github.com/deploymenttheory/terraform-provider-jamfpro/internal/common/plist"
 	sharedschemas "github.com/deploymenttheory/terraform-provider-jamfpro/internal/common/shared_schemas"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"howett.net/plist"
