+++ internal/services/macos_configuration_profile_plist_generator/resource_constructor.go
@@ -8,8 +8,8 @@ import (
 	"log"
 
 	"github.com/deploymenttheory/go-api-sdk-jamfpro/sdk/jamfpro"
-	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/common/configuration_profiles/plist"
 	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/common/constructors"
+	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/common/plist"
 	sharedschemas "github.com/deploymenttheory/terraform-provider-jamfpro/internal/common/shared_schemas"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
